refactored server so that express app is only created once in server-config file.
