+++ server/routes/auth.js
@@ -1,26 +1,7 @@
-var express = require('express');
 var crypto = require('crypto');
-var passport = require('passport');
 var SliceStrategy = require('passport-slice').Strategy;
-var db = require('../db/sequelize');
-var session = require('express-session');
-var cookieParser = require('cookie-parser');
-var bodyParser = require('body-parser');
-var morgan = require('morgan');
-var methodOverride = require('method-override');
-var helper = require('../utils/helper');
 
-require('dotenv').load();
-
-var app = express();
-  app.use(morgan('combined', {skip: function (req, res) {return res.statusCode < 400}}));
-  app.use(cookieParser());
-  app.use(bodyParser.urlencoded({extended: true}));
-  app.use(bodyParser.json());
-  app.use(methodOverride());
-  app.use(session({secret: process.env.SESSION_SECRET, cookie: {maxAge: 86400000}, resave: false, saveUninitialized: true}));
-  app.use(passport.initialize());
-  app.use(passport.session());
+module.exports = function(app, passport, db, helper) {
 
 // serialize and deserialize
 passport.serializeUser(function(userId, done) {
@@ -57,14 +38,14 @@ passport.use(new SliceStrategy({
 ));
 
 // Redirect the user to Slice for authentication.
-app.get('/slice', passport.authenticate('slice'));
+app.get('/auth/slice', passport.authenticate('slice'));
 
 // Slice has redirected the user back to the application.
 // Attempt to obtain an access token. If authorization granted,
 // the user is logged in. Otherwise, authentication has failed.
-app.get('/slice/callback', 
+app.get('/auth/slice/callback', 
   passport.authenticate('slice', {
     successRedirect: '/account', failureRedirect: '/login'
   }));
 
-module.exports = app;
+};
